split up core script. add eslint
